requirements.txt
# ============================================
# EYE-FONCIER — Dépendances Python
# ============================================

# Core Django
Django>=5.0,<5.2
psycopg2-binary>=2.9.9
gunicorn>=21.2.0

# Admin UI moderne
django-jazzmin>=3.0.0

# GeoDjango
#GDAL>=3.6.0

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-gis>=1.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
django-filter>=23.5

# Forms
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2

# Storage S3 (optionnel)
django-storages>=1.14.2
boto3>=1.34.0

# Cache Redis
django-redis>=5.4.0
redis>=5.0.1

# Images & PDF
Pillow>=10.2.0
reportlab>=4.1.0
PyPDF2>=3.0.1
xhtml2pdf>=0.2.15

# Géospatial
#fiona>=1.9.5
shapely>=2.0.2
pyproj>=3.6.0
rasterio>=1.3.9
geopandas>=0.14.0
numpy>=1.26.0
ezdxf>=1.1.0

# Data Science (Matching)
pandas>=2.1.0
scikit-learn>=1.4.0

# Sécurité
cryptography>=42.0.0
python-dotenv>=1.0.0

# Async Tasks
celery>=5.4.0
redis>=5.0.1

# WhatsApp (Twilio)
twilio>=9.0.0

# Dev & Utils
django-extensions>=3.2.3
django-debug-toolbar>=4.2.0
ipython>=8.20.0
